+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - `strict` type checking
+- `googleMapsApiKey` from forms app config used if google map api key had not been set in options
 
 ## [1.5.11] - 2024-05-21